+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Changes
 
 * [#233](https://github.com/rubocop-hq/rubocop-rails/pull/233): **(BREAKING)** Drop support for Ruby 2.3. ([@koic][])
+* [#236](https://github.com/rubocop-hq/rubocop-rails/pull/236): **(BREAKING)** Drop support for Rails 4.1 or lower. ([@koic][])
 
 ## 2.5.2 (2020-04-09)